lib/ukbitops: Move bitcount.h under lib namespace
This change moves bitcount.h under bitops/ in the default include path,
as is convention for secondary library headers.
Users of bitcount.h have been updated.
