+++ lib/ukbitops/include/uk/bitops/bitcount.h
